Add an optional environment variable for a qualifier
As the AWS documentation recommends, you should set an alias to allow for
control over what is run in production. Since that documentation uses 'prod' as
an example, that is what we should use by default.

This changes the default behavior, which would have been to invoke the $LATEST alias
prior to this change.